Clean-up, red tests
