+++ modules/http4s-swagger/test/src-ce2/TestCompat.scala
@@ -18,25 +18,14 @@ package smithy4s.http4s.swagger
 
 import cats.effect.Blocker
 import cats.effect.IO
-import smithy4s.HasId
-import smithy4s.ShapeId
 import weaver._
 
 import scala.concurrent.ExecutionContext
-import org.http4s.HttpRoutes
 
 trait TestCompat { self: BaseIOSuite with RunnableSuite[IO] =>
 
   def blocker: Blocker =
     Blocker.liftExecutionContext(ExecutionContext.Implicits.global)
 
-  def service = new HasId {
-    def id: ShapeId = ShapeId("foobar", "test-spec")
-  }
-
-  def docs(path: String) =
-    Docs.build[IO](blocker, path, swaggerUiPath = "swaggerui")(service)
-
-  def toApply(pad: PartiallyAppliedDocs): HttpRoutes[IO] =
-    pad.apply(blocker, service)
+  def mkDocs = smithy4s.http4s.swagger.docs[IO](blocker)
 }
