+++ modules/http4s-swagger/test/src-ce3/TestCompat.scala
@@ -17,20 +17,10 @@
 package smithy4s.http4s.swagger
 
 import cats.effect.IO
-import org.http4s._
-import smithy4s.HasId
-import smithy4s.ShapeId
 import weaver.BaseIOSuite
 
 trait TestCompat { self: BaseIOSuite =>
 
-  def service = new HasId {
-    def id: ShapeId = ShapeId("foobar", "test-spec")
-  }
+  def mkDocs = smithy4s.http4s.swagger.docs[IO]
 
-  def docs(path: String) =
-    Docs.build[IO](path, swaggerUiPath = "swaggerui")(service)
-
-  def toApply(pad: PartiallyAppliedDocs): HttpRoutes[IO] =
-    pad.apply(service)
 }
